Refactor database module into package
