+++ bot/db/__init__.py
@@ -1,0 +1,52 @@
+from .base import DB_PATH, init_db
+from .subjects import (
+    get_levels,
+    get_level_id_by_name,
+    get_term_id_by_name,
+    insert_level,
+    insert_term,
+    insert_subject,
+    get_terms_by_level,
+    get_subjects_by_level_and_term,
+    get_subject_id_by_name,
+    count_subjects,
+    term_feature_flags,
+    get_available_sections_for_subject,
+)
+from .materials import (
+    get_year_id_by_name,
+    get_lecturer_id_by_name,
+    insert_material,
+    insert_year,
+    insert_lecturer,
+    ensure_year_id,
+    ensure_lecturer_id,
+    get_years_for_subject_section,
+    get_lecturers_for_subject_section,
+    has_lecture_category,
+    list_lecture_titles,
+    list_lecture_titles_by_year,
+    list_lecture_titles_by_lecturer,
+    list_lecture_titles_by_lecturer_year,
+    get_years_for_subject_section_lecturer,
+    get_lecture_materials,
+    get_materials_by_category,
+    list_categories_for_subject_section_year,
+    list_categories_for_lecture,
+)
+
+__all__ = [
+    'DB_PATH', 'init_db',
+    'get_levels', 'get_level_id_by_name', 'get_term_id_by_name',
+    'insert_level', 'insert_term', 'insert_subject',
+    'get_terms_by_level', 'get_subjects_by_level_and_term', 'get_subject_id_by_name',
+    'count_subjects', 'term_feature_flags', 'get_available_sections_for_subject',
+    'get_year_id_by_name', 'get_lecturer_id_by_name', 'insert_material',
+    'insert_year', 'insert_lecturer', 'ensure_year_id', 'ensure_lecturer_id',
+    'get_years_for_subject_section', 'get_lecturers_for_subject_section',
+    'has_lecture_category', 'list_lecture_titles', 'list_lecture_titles_by_year',
+    'list_lecture_titles_by_lecturer', 'list_lecture_titles_by_lecturer_year',
+    'get_years_for_subject_section_lecturer', 'get_lecture_materials',
+    'get_materials_by_category', 'list_categories_for_subject_section_year',
+    'list_categories_for_lecture',
+]
