+++ bot/db/materials.py
@@ -1,55 +1,17 @@
-# db.py
-# طبقة الوصول لقاعدة البيانات (SQLite عبر aiosqlite)
-# تضم فقط الدوال التي يستخدمها bot.py في الإصدار الحالي.
-
-import os
 import aiosqlite
 
-DB_PATH = "database/archive.db"
+from .base import DB_PATH
 
 
 # -----------------------------------------------------------------------------
-# تهيئة قاعدة البيانات
+# Helpers for years/lecturers
 # -----------------------------------------------------------------------------
-async def init_db() -> None:
-    """
-    يضمن وجود مجلد قاعدة البيانات، ثم ينفّذ ملف schema/init.sql مرة واحدة.
-    """
-    os.makedirs(os.path.dirname(DB_PATH), exist_ok=True)
-    async with aiosqlite.connect(DB_PATH) as db:
-        with open("database/init.sql", "r", encoding="utf-8") as f:
-            await db.executescript(f.read())
-        await db.commit()
-
-
-# -----------------------------------------------------------------------------
-# قراءات أساسية (مستويات / أترام / مواد)
-# -----------------------------------------------------------------------------
-async def get_levels():
-    """
-    يرجع قائمة المستويات بشكل [(id, name), ...] مرتبة بالمعرّف.
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute("SELECT id, name FROM levels ORDER BY id")
-        return await cur.fetchall()
-
-async def get_level_id_by_name(name: str) -> int | None:
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute("SELECT id FROM levels WHERE name=?", (name,))
-        row = await cur.fetchone()
-        return row[0] if row else None
-
-async def get_term_id_by_name(name: str) -> int | None:
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute("SELECT id FROM terms WHERE name=?", (name,))
-        row = await cur.fetchone()
-        return row[0] if row else None
-
 async def get_year_id_by_name(name: str) -> int | None:
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute("SELECT id FROM years WHERE name=?", (name,))
         row = await cur.fetchone()
         return row[0] if row else None
+
 
 async def get_lecturer_id_by_name(name: str) -> int | None:
     async with aiosqlite.connect(DB_PATH) as db:
@@ -57,33 +19,15 @@ async def get_lecturer_id_by_name(name: str) -> int | None:
         row = await cur.fetchone()
         return row[0] if row else None
 
-async def insert_level(name: str):
-    async with aiosqlite.connect(DB_PATH) as db:
-        await db.execute("INSERT OR IGNORE INTO levels (name) VALUES (?)", (name,))
-        await db.commit()
 
-async def insert_term(name: str):
-    async with aiosqlite.connect(DB_PATH) as db:
-        await db.execute("INSERT OR IGNORE INTO terms (name) VALUES (?)", (name,))
-        await db.commit()
-
-async def insert_subject(code: str, name: str, level_id: int, term_id: int):
-    async with aiosqlite.connect(DB_PATH) as db:
-        await db.execute(
-            "INSERT INTO subjects (code, name, level_id, term_id) VALUES (?, ?, ?, ?)",
-            (code, name, level_id, term_id)
-        )
-        await db.commit()
-
-# ---------- materials ----------
 async def insert_material(
     subject_id: int,
-    section: str,       # 'theory' | 'discussion' | 'lab' | 'syllabus' | 'apps'
-    category: str,      # 'lecture'|'exam'|'booklet'|'board_images'|'video'|'simulation'|'summary'|'notes'|'external_link'
+    section: str,
+    category: str,
     title: str,
     url: str | None = None,
     year_id: int | None = None,
-    lecturer_id: int | None = None
+    lecturer_id: int | None = None,
 ):
     async with aiosqlite.connect(DB_PATH) as db:
         await db.execute(
@@ -91,7 +35,7 @@ async def insert_material(
             INSERT INTO materials (subject_id, section, category, title, url, year_id, lecturer_id)
             VALUES (?, ?, ?, ?, ?, ?, ?)
             """,
-            (subject_id, section, category, title, url, year_id, lecturer_id)
+            (subject_id, section, category, title, url, year_id, lecturer_id),
         )
         await db.commit()
 
@@ -101,13 +45,15 @@ async def insert_year(name: str):
         await db.execute("INSERT OR IGNORE INTO years (name) VALUES (?)", (name,))
         await db.commit()
 
+
 async def insert_lecturer(name: str, role: str = "lecturer"):
     async with aiosqlite.connect(DB_PATH) as db:
         await db.execute(
             "INSERT OR IGNORE INTO lecturers (name, role) VALUES (?, ?)",
-            (name, role)
+            (name, role),
         )
         await db.commit()
+
 
 async def ensure_year_id(name: str) -> int:
     _id = await get_year_id_by_name(name)
@@ -119,6 +65,7 @@ async def ensure_year_id(name: str) -> int:
         raise RuntimeError(f"Failed to create year: {name}")
     return _id
 
+
 async def ensure_lecturer_id(name: str, role: str = "lecturer") -> int:
     _id = await get_lecturer_id_by_name(name)
     if _id is not None:
@@ -128,131 +75,10 @@ async def ensure_lecturer_id(name: str, role: str = "lecturer") -> int:
     if _id is None:
         raise RuntimeError(f"Failed to create lecturer: {name}")
     return _id
-
-
-async def get_terms_by_level(level_id: int):
-    """
-    يرجع قائمة الأترام المرتبطة بالمستوى المحدد: [(term_id, term_name), ...]
-    يعتمد على وجود مواد (subjects) مرتبطة بذلك المستوى.
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute(
-            """
-            SELECT DISTINCT t.id, t.name
-            FROM terms t
-            JOIN subjects s ON s.term_id = t.id
-            WHERE s.level_id = ?
-            ORDER BY t.id
-            """,
-            (level_id,),
-        )
-        return await cur.fetchall()
-
-
-async def get_subjects_by_level_and_term(level_id: int, term_id: int):
-    """
-    يرجع أسماء المواد لهذا (المستوى، الترم) بالشكل [(name,), ...]
-    (يُستخدم في بناء أزرار اختيار المادة).
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute(
-            "SELECT name FROM subjects WHERE level_id = ? AND term_id = ? ORDER BY id",
-            (level_id, term_id),
-        )
-        return await cur.fetchall()
-
-
-async def get_subject_id_by_name(level_id: int, term_id: int, subject_name: str) -> int | None:
-    """
-    يرجع معرّف المادة بحسب الاسم والمستوى والترم، أو None إذا لم توجد.
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute(
-            "SELECT id FROM subjects WHERE level_id=? AND term_id=? AND name=?",
-            (level_id, term_id, subject_name),
-        )
-        row = await cur.fetchone()
-        return row[0] if row else None
-
-
 # -----------------------------------------------------------------------------
-# خصائص ديناميكية لبناء القوائم
-# -----------------------------------------------------------------------------
-async def count_subjects(level_id: int, term_id: int) -> int:
-    """
-    عدد المواد في (مستوى/ترم) — يُستخدم لاستخراج وجود مواد من عدمه.
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute(
-            "SELECT COUNT(*) FROM subjects WHERE level_id=? AND term_id=?",
-            (level_id, term_id),
-        )
-        (n,) = await cur.fetchone()
-        return n
-
-
-async def term_feature_flags(level_id: int, term_id: int) -> dict:
-    """
-    يُعيد أعلام توضح ما إذا كانت هناك مواد/سيلابس/روابط خارجية في الترم:
-    {
-        "has_subjects": bool,
-        "has_syllabus": bool,
-        "has_links": bool
-    }
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        # سيلابس
-        cur = await db.execute(
-            """
-            SELECT 1
-            FROM materials m
-            JOIN subjects s ON s.id = m.subject_id
-            WHERE s.level_id=? AND s.term_id=? AND m.section='syllabus'
-            LIMIT 1
-            """,
-            (level_id, term_id),
-        )
-        has_syllabus = (await cur.fetchone()) is not None
-
-        # روابط خارجية
-        cur = await db.execute(
-            """
-            SELECT 1
-            FROM materials m
-            JOIN subjects s ON s.id = m.subject_id
-            WHERE s.level_id=? AND s.term_id=? AND m.category='external_link'
-            LIMIT 1
-            """,
-            (level_id, term_id),
-        )
-        has_links = (await cur.fetchone()) is not None
-
-        n_subj = await count_subjects(level_id, term_id)
-
-    return {"has_subjects": n_subj > 0, "has_syllabus": has_syllabus, "has_links": has_links}
-
-
-async def get_available_sections_for_subject(subject_id: int) -> list[str]:
-    """
-    الأقسام المتوفرة فعليًا لهذه المادة من جدول materials
-    (مثال: theory / discussion / lab / syllabus / apps).
-    """
-    async with aiosqlite.connect(DB_PATH) as db:
-        cur = await db.execute(
-            "SELECT DISTINCT section FROM materials WHERE subject_id=?",
-            (subject_id,),
-        )
-        rows = await cur.fetchall()
-        return [r[0] for r in rows]
-
-
-# -----------------------------------------------------------------------------
-# فلاتر حسب السنة/المحاضر/التصنيفات
+# Filters for years/lecturers/categories
 # -----------------------------------------------------------------------------
 async def get_years_for_subject_section(subject_id: int, section: str):
-    """
-    السنوات المتاحة لمادة + قسم بالشكل [(year_id, year_name), ...] تنازليًا بالاسم.
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -268,9 +94,6 @@ async def get_years_for_subject_section(subject_id: int, section: str):
 
 
 async def get_lecturers_for_subject_section(subject_id: int, section: str):
-    """
-    المحاضرون المتاحون لمادة + قسم بالشكل [(lect_id, lect_name), ...].
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -286,9 +109,6 @@ async def get_lecturers_for_subject_section(subject_id: int, section: str):
 
 
 async def has_lecture_category(subject_id: int, section: str) -> bool:
-    """
-    هل توجد محاضرات (category='lecture') لهذه المادة + القسم؟
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -303,9 +123,6 @@ async def has_lecture_category(subject_id: int, section: str) -> bool:
 
 
 async def list_lecture_titles(subject_id: int, section: str) -> list[str]:
-    """
-    عناوين المحاضرات لجميع السنوات/المحاضرين (distinct) مرتبة أبجديًا.
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -320,9 +137,6 @@ async def list_lecture_titles(subject_id: int, section: str) -> list[str]:
 
 
 async def list_lecture_titles_by_year(subject_id: int, section: str, year_id: int) -> list[str]:
-    """
-    عناوين المحاضرات لسنة محددة.
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -337,9 +151,6 @@ async def list_lecture_titles_by_year(subject_id: int, section: str, year_id: in
 
 
 async def list_lecture_titles_by_lecturer(subject_id: int, section: str, lecturer_id: int) -> list[str]:
-    """
-    عناوين المحاضرات لمحاضر محدد.
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -356,9 +167,6 @@ async def list_lecture_titles_by_lecturer(subject_id: int, section: str, lecture
 async def list_lecture_titles_by_lecturer_year(
     subject_id: int, section: str, lecturer_id: int, year_id: int
 ) -> list[str]:
-    """
-    عناوين المحاضرات لمحاضر + سنة معًا.
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -373,9 +181,6 @@ async def list_lecture_titles_by_lecturer_year(
 
 
 async def get_years_for_subject_section_lecturer(subject_id: int, section: str, lecturer_id: int):
-    """
-    السنوات المتاحة لمحاضر معيّن داخل مادة + قسم.
-    """
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(
             """
@@ -389,10 +194,8 @@ async def get_years_for_subject_section_lecturer(subject_id: int, section: str, 
             (subject_id, section, lecturer_id),
         )
         return await cur.fetchall()
-
-
 # -----------------------------------------------------------------------------
-# جلب المواد/التصنيفات
+# Fetch materials
 # -----------------------------------------------------------------------------
 async def get_lecture_materials(
     subject_id: int,
@@ -402,10 +205,6 @@ async def get_lecture_materials(
     lecturer_id: int | None = None,
     title: str | None = None,
 ):
-    """
-    يرجع ملفات (category='lecture') مع إمكانية التصفية بالسنة/المحاضر/العنوان.
-    الشكل: [(id, title, url), ...]
-    """
     q = """
         SELECT id, title, url
         FROM materials
@@ -437,11 +236,6 @@ async def get_materials_by_category(
     lecturer_id: int | None = None,
     title: str | None = None,
 ):
-    """
-    يرجع مواد حسب تصنيف عام (امتحانات/ملازم/ملخصات/… إلخ)
-    مع مرشحات اختيارية (سنة/محاضر/عنوان).
-    الشكل: [(id, title, url), ...]
-    """
     q = """
         SELECT id, title, url
         FROM materials
@@ -470,11 +264,6 @@ async def list_categories_for_subject_section_year(
     year_id: int,
     lecturer_id: int | None = None,
 ) -> list[str]:
-    """
-    تصنيفات سنة متاحة لمادة + قسم + سنة (باستثناء:
-    - 'lecture' (المحاضرات تُعرض من زر منفصل)
-    - مرفقات المحاضرات التي يجب ألا تظهر على شاشة السنة)
-    """
     lecture_attachment_cats = (
         "slides",
         "audio",
@@ -511,9 +300,6 @@ async def list_categories_for_lecture(
     year_id: int | None = None,
     lecturer_id: int | None = None,
 ) -> list[str]:
-    """
-    التصنيفات المتوفرة داخل محاضرة محددة (lecture/slides/audio/…).
-    """
     q = """
         SELECT DISTINCT category
         FROM materials
@@ -530,7 +316,3 @@ async def list_categories_for_lecture(
     async with aiosqlite.connect(DB_PATH) as db:
         cur = await db.execute(q, tuple(params))
         return [r[0] for r in await cur.fetchall()]
-
-
-# __________الاصدار الثاني ____________
-
